Indexer
Traverse through a directory reading all files and
subdirectories inside of it using the built in library FTW(). During the
traversal we read contents of the files in the directory and tokenize
the
contents of the file using the tokenize() function and then indexing
tokens in
ascending order keeping track of how many times a specific token was
used in each file it exists inside and then storing and sorting inside
of a
hashtable.